+++ lib/map/empty_map.dart
@@ -1,0 +1,6 @@
+import 'package:bonfire/map/world_map.dart';
+
+class EmptyWorldMap extends WorldMap {
+  EmptyWorldMap({double tileSizeToUpdate = 0})
+      : super([], tileSizeToUpdate: tileSizeToUpdate);
+}